pyproject.toml
[tool.poetry]
name = "rag-intro"
version = "0.1.0"
description = ""
authors = ["Mario Toribio <yuniormtr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.2"
langchain-openai = "^0.1.3"
langchain = "^0.1.16"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
pinecone-client = "^3.2.2"
langchain-cli = "^0.0.21"
langchain-pinecone = "^0.1.0"
ipykernel = "^6.29.4"
langchain-google-genai = "^1.0.2"
streamlit = "^1.33.0"
st-annotated-text = "^4.0.1"
streamlit-javascript = "^0.1.5"
streamlit-js-eval = "^0.1.7"
unidecode = "^1.3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
